# public/swagger.json
{
	"components": {
...
		"schemas": {
			"Types.ObjectId": {
				"type": "string"
			},
			"IUsers": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"type": "string"
							}
						]
					},
					"email": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"profilePicture": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"gender": {
						"type": "string"
					},
					"DOB": {
						"type": "string",
						"format": "date-time"
					},
					"address": {
						"items": {
							"properties": {
								"manipolicity": {
									"type": "string"
								},
								"government": {
									"type": "string"
								},
								"country": {
									"type": "string"
								}
							},
							"required": [
								"manipolicity",
								"government",
								"country"
							],
							"type": "object"
						},
						"type": "array"
					},
					"verified": {
						"type": "string",
						"enum": [
							"notSent",
							"pending",
							"verified"
						]
					},
					"status": {
						"type": "string",
						"enum": [
							"inActive",
							"active",
							"suspended",
							"lost",
							"deleted"
						]
					},
					"accountType": {
						"type": "string",
						"enum": [
							"PT",
							"EM",
							"PA"
						]
					},
					"lastLoginDate": {
						"type": "string",
						"format": "date-time"
					},
					"accountSetting": {
						"additionalProperties": false,
						"type": "object"
					},
					"languages": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"maritalStatus": {
						"type": "string",
						"enum": [
							"married",
							"single",
							"divorced",
							"widow"
						]
					},
					"url": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"phoneNumber",
					"password",
					"profilePicture",
					"firstName",
					"lastName",
					"gender",
					"DOB",
					"verified",
					"status",
					"accountType",
					"lastLoginDate",
					"accountSetting",
					"languages",
					"maritalStatus"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IServiceProviders": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"type": "string"
							}
						]
					},
					"uID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"$ref": "#/components/schemas/IUsers"
							}
						]
					},
					"bio": {
						"type": "string"
					},
					"specialties": {
						"type": "string"
					},
					"preferredServiceType": {
						"anyOf": [
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							{
								"items": {
									"type": "string",
									"enum": [
										"online",
										"home",
										"office"
									]
								},
								"type": "array"
							}
						]
					},
					"minSessionFee": {
						"type": "number",
						"format": "double"
					},
					"maxSessionFee": {
						"type": "number",
						"format": "double"
					},
					"documents": {
						"properties": {
							"attType": {
								"type": "string"
							},
							"name": {
								"type": "string"
							},
							"url": {
								"type": "string"
							}
						},
						"required": [
							"attType",
							"name",
							"url"
						],
						"type": "object"
					},
					"reviewerUIDs": {
						"anyOf": [
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/Types.ObjectId"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/IUsers"
								},
								"type": "array"
							}
						]
					},
					"verificationStatus": {
						"type": "string",
						"enum": [
							"notSubmitted",
							"pendingReview",
							"inReview",
							"verified",
							"rejected"
						]
					},
					"verificationDate": {
						"type": "string",
						"format": "date-time"
					},
					"verifiedByUID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"$ref": "#/components/schemas/IUsers"
							}
						]
					},
					"url": {
						"type": "string"
					}
				},
				"required": [
					"bio",
					"specialties",
					"preferredServiceType",
					"minSessionFee",
					"maxSessionFee",
					"verificationStatus",
					"verificationDate"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IClients": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							}
						]
					},
					"uID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"$ref": "#/components/schemas/IUsers"
							}
						]
					},
					"preferredServiceType": {
						"anyOf": [
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							{
								"items": {
									"type": "string",
									"enum": [
										"online",
										"home",
										"office"
									]
								},
								"type": "array"
							}
						]
					},
					"diseases": {
						"type": "string"
					},
					"preferences": {
						"additionalProperties": false,
						"type": "object"
					},
					"url": {
						"type": "string"
					}
				},
				"required": [
					"preferredServiceType",
					"diseases",
					"preferences"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ISessions": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"type": "string"
							}
						]
					},
					"sessionType": {
						"type": "string",
						"enum": [
							"group",
							"individual"
						]
					},
					"serviceProvidersID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"$ref": "#/components/schemas/IServiceProviders"
							}
						]
					},
					"clientsIDs": {
						"anyOf": [
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/Types.ObjectId"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/IClients"
								},
								"type": "array"
							}
						]
					},
					"name": {
						"type": "string"
					},
					"details": {
						"type": "string"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"duration": {
						"type": "number",
						"format": "double"
					},
					"serviceType": {
						"type": "string",
						"enum": [
							"Online",
							"Home",
							"Office"
						]
					},
					"location": {
						"additionalProperties": false,
						"type": "object"
					},
					"attachments": {
						"properties": {
							"attachmentType": {
								"type": "string"
							},
							"attachmentName": {
								"type": "string"
							},
							"attachmentUrl": {
								"type": "string"
							}
						},
						"required": [
							"attachmentType",
							"attachmentName",
							"attachmentUrl"
						],
						"type": "object"
					},
					"requirements": {
						"type": "string"
					},
					"ratings": {
						"items": {
							"properties": {
								"ratingDate": {
									"type": "string",
									"format": "date-time"
								},
								"ratingValue": {
									"type": "string"
								},
								"raterUID": {
									"anyOf": [
										{
											"type": "string"
										},
										{
											"$ref": "#/components/schemas/Types.ObjectId"
										},
										{
											"$ref": "#/components/schemas/IUsers"
										}
									]
								}
							},
							"required": [
								"ratingDate",
								"ratingValue",
								"raterUID"
							],
							"type": "object"
						},
						"type": "array"
					},
					"reviews": {
						"items": {
							"properties": {
								"reviewDate": {
									"type": "string",
									"format": "date-time"
								},
								"reviewDetails": {
									"type": "string"
								},
								"reviewerUID": {
									"anyOf": [
										{
											"type": "string"
										},
										{
											"$ref": "#/components/schemas/Types.ObjectId"
										},
										{
											"$ref": "#/components/schemas/IUsers"
										}
									]
								}
							},
							"required": [
								"reviewDate",
								"reviewDetails",
								"reviewerUID"
							],
							"type": "object"
						},
						"type": "array"
					},
					"sessionFee": {
						"type": "number",
						"format": "double"
					},
					"payments": {
						"properties": {
							"amount": {
								"type": "number",
								"format": "double"
							},
							"payerID": {
								"anyOf": [
									{
										"type": "string"
									},
									{
										"$ref": "#/components/schemas/Types.ObjectId"
									},
									{
										"$ref": "#/components/schemas/IUsers"
									}
								]
							},
							"paymentMethod": {
								"type": "string"
							},
							"discount": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"amount",
							"payerID",
							"paymentMethod",
							"discount"
						],
						"type": "object"
					},
					"status": {
						"type": "string",
						"enum": [
							"initiated",
							"agreed",
							"canceled",
							"finished"
						]
					},
					"doctorReferral": {
						"type": "string"
					},
					"url": {
						"type": "string"
					}
				},
				"required": [
					"sessionType",
					"name",
					"details",
					"startDate",
					"duration",
					"serviceType",
					"location",
					"requirements",
					"sessionFee",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IDisputes": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"type": "string"
							}
						]
					},
					"sessionID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"$ref": "#/components/schemas/ISessions"
							}
						]
					},
					"firstPartyUID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"$ref": "#/components/schemas/IUsers"
							}
						]
					},
					"secondUID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"$ref": "#/components/schemas/IUsers"
							}
						]
					},
					"topic": {
						"type": "string"
					},
					"details": {
						"type": "string"
					},
					"attachments": {
						"items": {
							"properties": {
								"disType": {
									"type": "string"
								},
								"name": {
									"type": "string"
								},
								"url": {
									"type": "string"
								}
							},
							"required": [
								"disType",
								"name",
								"url"
							],
							"type": "object"
						},
						"type": "array"
					},
					"status": {
						"type": "string",
						"enum": [
							"sent",
							"received",
							"in-progress",
							"suspended",
							"rejected",
							"resolved"
						]
					},
					"resolverUID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"$ref": "#/components/schemas/IUsers"
							}
						]
					},
					"inProgressDate": {
						"type": "string",
						"format": "date-time"
					},
					"receivedDate": {
						"type": "string",
						"format": "date-time"
					},
					"suspendedDate": {
						"type": "string",
						"format": "date-time"
					},
					"closedDate": {
						"type": "string",
						"format": "date-time"
					},
					"url": {
						"type": "string"
					}
				},
				"required": [
					"topic",
					"details",
					"status",
					"inProgressDate",
					"receivedDate",
					"suspendedDate",
					"closedDate"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IAgreements": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							}
						]
					},
					"name": {
						"type": "string"
					},
					"parties": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"endDate": {
						"type": "string",
						"format": "date-time"
					},
					"details": {
						"type": "string"
					},
					"attachments": {
						"items": {
							"additionalProperties": false,
							"type": "object"
						},
						"type": "array"
					},
					"reminder": {
						"type": "boolean"
					},
					"url": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"parties",
					"startDate",
					"endDate",
					"details",
					"reminder"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICommunications": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							}
						]
					},
					"referenceType": {
						"type": "string",
						"enum": [
							"Agreement",
							"Session",
							"Dispute",
							"User"
						]
					},
					"referenceID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"$ref": "#/components/schemas/ISessions"
							},
							{
								"$ref": "#/components/schemas/IAgreements"
							},
							{
								"$ref": "#/components/schemas/IDisputes"
							},
							{
								"$ref": "#/components/schemas/IUsers"
							}
						]
					},
					"partiesUIDs": {
						"anyOf": [
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/Types.ObjectId"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/IUsers"
								},
								"type": "array"
							}
						]
					},
					"lastUpdate": {
						"type": "string",
						"format": "date-time"
					},
					"messages": {
						"items": {
							"properties": {
								"deliveryDetails": {
									"items": {
										"additionalProperties": false,
										"type": "object"
									},
									"type": "array"
								},
								"sendDate": {
									"type": "string",
									"format": "date-time"
								},
								"senderUID": {
									"anyOf": [
										{
											"type": "string"
										},
										{
											"$ref": "#/components/schemas/Types.ObjectId"
										},
										{
											"$ref": "#/components/schemas/IUsers"
										}
									]
								},
								"messageContent": {
									"type": "string"
								},
								"messageType": {
									"type": "string"
								}
							},
							"required": [
								"deliveryDetails",
								"sendDate",
								"senderUID",
								"messageContent",
								"messageType"
							],
							"type": "object"
						},
						"type": "array"
					},
					"url": {
						"type": "string"
					}
				},
				"required": [
					"referenceType",
					"lastUpdate",
					"messages"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IRoles": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							}
						]
					},
					"name": {
						"type": "string"
					},
					"employees": {
						"anyOf": [
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/Types.ObjectId"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/IEmployees"
								},
								"type": "array"
							}
						]
					},
					"users": {
						"anyOf": [
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/Types.ObjectId"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/IUsers"
								},
								"type": "array"
							}
						]
					},
					"service_provider": {
						"anyOf": [
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/Types.ObjectId"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/IServiceProviders"
								},
								"type": "array"
							}
						]
					},
					"clients": {
						"anyOf": [
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/Types.ObjectId"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/IClients"
								},
								"type": "array"
							}
						]
					},
					"sessions": {
						"anyOf": [
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/Types.ObjectId"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/ISessions"
								},
								"type": "array"
							}
						]
					},
					"communications": {
						"anyOf": [
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/Types.ObjectId"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/ICommunications"
								},
								"type": "array"
							}
						]
					},
					"disputes": {
						"anyOf": [
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/Types.ObjectId"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/IDisputes"
								},
								"type": "array"
							}
						]
					},
					"enum_values": {
						"anyOf": [
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/Types.ObjectId"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/IEnumValues"
								},
								"type": "array"
							}
						]
					},
					"url": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IEmployees": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							}
						]
					},
					"uID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"$ref": "#/components/schemas/IUsers"
							}
						]
					},
					"roleID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"$ref": "#/components/schemas/IRoles"
							}
						]
					},
					"salary": {
						"type": "number",
						"format": "double"
					},
					"attachments": {
						"items": {
							"properties": {
								"empType": {
									"type": "string"
								},
								"url": {
									"type": "string"
								},
								"name": {
									"type": "string"
								}
							},
							"required": [
								"empType",
								"url",
								"name"
							],
							"type": "object"
						},
						"type": "array"
					},
					"url": {
						"type": "string"
					}
				},
				"required": [
					"salary"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IEnumValues": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"type": "string"
							}
						]
					},
					"enumName": {
						"type": "string"
					},
					"enumValues": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"enumNote": {
						"type": "string"
					},
					"url": {
						"type": "string"
					}
				},
				"required": [
					"enumName",
					"enumValues",
					"enumNote"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_IRoles_": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							}
						]
					},
					"name": {
						"type": "string"
					},
					"employees": {
						"anyOf": [
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/Types.ObjectId"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/IEmployees"
								},
								"type": "array"
							}
						]
					},
					"users": {
						"anyOf": [
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/Types.ObjectId"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/IUsers"
								},
								"type": "array"
							}
						]
					},
					"service_provider": {
						"anyOf": [
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/Types.ObjectId"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/IServiceProviders"
								},
								"type": "array"
							}
						]
					},
					"clients": {
						"anyOf": [
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/Types.ObjectId"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/IClients"
								},
								"type": "array"
							}
						]
					},
					"sessions": {
						"anyOf": [
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/Types.ObjectId"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/ISessions"
								},
								"type": "array"
							}
						]
					},
					"communications": {
						"anyOf": [
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/Types.ObjectId"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/ICommunications"
								},
								"type": "array"
							}
						]
					},
					"disputes": {
						"anyOf": [
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/Types.ObjectId"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/IDisputes"
								},
								"type": "array"
							}
						]
					},
					"enum_values": {
						"anyOf": [
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/Types.ObjectId"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/IEnumValues"
								},
								"type": "array"
							}
						]
					},
					"url": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/roles/{roleId}": {
			"get": {
				"operationId": "GetRole",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IRoles"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "The requested role is not found"
					}
				},
				"description": "Get role details",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "roleId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "631312818d796320172e58e5"
					}
				]
			},
			"delete": {
				"operationId": "DeleteRole",
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IRoles"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "The requested role is not found"
					}
				},
				"description": "Delete role",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "roleId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "631312818d796320172e58e5"
					}
				]
			}
		},
		"/roles/create": {
			"post": {
				"operationId": "CreateRole",
				"responses": {
					"200": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IRoles"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "Admin",
											"users": [
												"63124dffdcf1e4974079a441"
											]
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation failed"
					}
				},
				"description": "Create role",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IRoles"
							}
						}
					}
				}
			}
		},
		"/roles/update/{roleId}": {
			"put": {
				"operationId": "UpdateRole",
				"responses": {
					"200": {
						"description": "Updated",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IRoles"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"description": "Update role",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "roleId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "631312818d796320172e58e5"
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_IRoles_"
							},
							"example": "631312818d796320172e58e5"
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}